feature: camera animation added and return to user position included with back handler

+++ app/src/main/java/com/example/technicaltestmaps/presentation/composables/MapView.kt
@@ -3,10 +3,12 @@ package com.example.technicaltestmaps.presentation.composables
 import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.Canvas
+import android.util.Log
 import androidx.annotation.DrawableRes
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.ui.Modifier
@@ -17,6 +19,7 @@ import androidx.core.graphics.createBitmap
 import com.example.technicaltestmaps.R
 import com.example.technicaltestmaps.domain.model.FavoritePoint
 import com.example.technicaltestmaps.domain.model.PointType
+import com.mapbox.android.gestures.MoveGestureDetector
 import com.mapbox.geojson.FeatureCollection
 import com.mapbox.geojson.Point
 import com.mapbox.maps.CameraOptions
@@ -26,10 +29,14 @@ import com.mapbox.maps.extension.style.layers.addLayer
 import com.mapbox.maps.extension.style.layers.generated.fillLayer
 import com.mapbox.maps.extension.style.sources.addSource
 import com.mapbox.maps.extension.style.sources.generated.geoJsonSource
+import com.mapbox.maps.plugin.animation.MapAnimationOptions
+import com.mapbox.maps.plugin.animation.easeTo
 import com.mapbox.maps.plugin.annotation.annotations
 import com.mapbox.maps.plugin.annotation.generated.PointAnnotationOptions
 import com.mapbox.maps.plugin.annotation.generated.createPointAnnotationManager
+import com.mapbox.maps.plugin.gestures.OnMoveListener
 import com.mapbox.maps.plugin.gestures.gestures
+import timber.log.Timber
 
 @Composable
 fun MapView(
@@ -38,7 +45,9 @@ fun MapView(
     userLocation: Point?,
     selectedPoint: FavoritePoint?,
     featureCollection: FeatureCollection?,
-    onMapLongClick: (Point) -> Unit
+    onMapLongClick: (Point) -> Unit,
+    shouldCenterOnUser: Boolean,
+    onUserCentered: () -> Unit
 ) {
     val context = LocalContext.current
     val mapView = remember { MapView(context) }
@@ -52,15 +61,17 @@ fun MapView(
     val currentSelectedPoint by rememberUpdatedState(selectedPoint)
     val currentUserLocation by rememberUpdatedState(userLocation)
 
+    val styleLoaded = remember { mutableStateOf(false) }
+    val hasUserMovedMap = remember { mutableStateOf(false) }
+
     AndroidView(
         modifier = modifier,
         factory = { mapView },
         update = { view ->
             val mapboxMap = view.getMapboxMap()
 
-            mapboxMap.loadStyleUri(Style.MAPBOX_STREETS) { style ->
-                val safeFeatureCollection =
-                    featureCollection ?: FeatureCollection.fromFeatures(emptyList())
+            mapboxMap.loadStyleUri(Style.OUTDOORS) { style ->
+                val safeFeatureCollection = featureCollection ?: FeatureCollection.fromFeatures(emptyList())
 
                 val source = geoJsonSource("geojson-source") {
                     featureCollection(safeFeatureCollection)
@@ -73,15 +84,12 @@ fun MapView(
                 }
                 style.addLayer(layer)
 
-                val target = selectedPoint?.let {
-                    Point.fromLngLat(it.longitude, it.latitude)
-                } ?: userLocation
-
-                target?.let {
+                // Centrar la cámara si hay ubicación del usuario
+                userLocation?.let {
                     mapboxMap.setCamera(
                         CameraOptions.Builder()
                             .center(it)
-                            .zoom(16.0)
+                            .zoom(12.0)
                             .build()
                     )
                 }
@@ -91,44 +99,56 @@ fun MapView(
                     onMapLongClick(point)
                     true
                 }
+                gesturesPlugin.addOnMoveListener(object : OnMoveListener {
+                    override fun onMoveBegin(detector: MoveGestureDetector) {
+                        hasUserMovedMap.value = true
+                    }
+
+                    override fun onMove(detector: MoveGestureDetector): Boolean = false
+
+                    override fun onMoveEnd(detector: MoveGestureDetector) {}
+                })
+
+                styleLoaded.value = true
             }
         }
     )
 
-    LaunchedEffect(points, currentSelectedPoint, currentUserLocation) {
+    LaunchedEffect(currentUserLocation, styleLoaded.value) {
+        if (styleLoaded.value && currentUserLocation != null && !hasUserMovedMap.value) {
+            mapView.getMapboxMap().setCamera(
+                CameraOptions.Builder()
+                    .center(currentUserLocation)
+                    .zoom(12.0)
+                    .build()
+            )
+        }
+    }
+
+    LaunchedEffect(points, currentSelectedPoint, currentUserLocation, styleLoaded.value) {
+        if (!styleLoaded.value) return@LaunchedEffect
+
         mapView.getMapboxMap().getStyle()?.let {
             annotationManager.deleteAll()
             viewAnnotationManager.removeAllViewAnnotations()
 
             points.forEach { fav ->
                 val point = Point.fromLngLat(fav.longitude, fav.latitude)
+                val iconRes = if (fav.type == PointType.ALERT)
+                    R.drawable.ic_alert_map_location_icon
+                else
+                    R.drawable.ic_red_marker
 
-                if (fav.type == PointType.ALERT) {
-                    val bitmap = bitmapFromDrawableRes(
-                        context,
-                        R.drawable.ic_alert_map_location_icon
-                    )
-                    bitmap?.let { bmp ->
-                        val options = PointAnnotationOptions()
-                            .withPoint(point)
-                            .withIconImage(bmp)
-                        annotationManager.create(options)
-                    }
-                } else {
-                    val bitmap = bitmapFromDrawableRes(
-                        context,
-                        R.drawable.ic_red_marker
-                    )
-                    bitmap?.let { bmp ->
-                        val options = PointAnnotationOptions()
-                            .withPoint(point)
-                            .withIconImage(bmp)
-                        annotationManager.create(options)
-                    }
+                bitmapFromDrawableRes(context, iconRes)?.let { bmp ->
+                    val options = PointAnnotationOptions()
+                        .withPoint(point)
+                        .withIconImage(bmp)
+                    annotationManager.create(options)
                 }
             }
 
             currentUserLocation?.let {
+                Timber.tag("MapView").d("User location: $currentUserLocation")
                 val bitmap = bitmapFromDrawableRes(context, R.drawable.ic_blue_marker)
                 bitmap?.let { bmp ->
                     val options = PointAnnotationOptions()
@@ -137,6 +157,34 @@ fun MapView(
                     annotationManager.create(options)
                 }
             }
+        }
+    }
+
+    LaunchedEffect(currentSelectedPoint, styleLoaded.value) {
+        if (styleLoaded.value && currentSelectedPoint != null) {
+            val mapboxMap = mapView.getMapboxMap()
+            val target = Point.fromLngLat(currentSelectedPoint!!.longitude, currentSelectedPoint!!.latitude)
+            mapboxMap.easeTo(
+                CameraOptions.Builder()
+                    .center(target)
+                    .zoom(14.0)
+                    .build(),
+                MapAnimationOptions.mapAnimationOptions {
+                    duration(1000L)
+                }
+            )
+        }
+    }
+
+    LaunchedEffect(shouldCenterOnUser, currentUserLocation, styleLoaded.value) {
+        if (shouldCenterOnUser && styleLoaded.value && currentUserLocation != null) {
+            mapView.getMapboxMap().setCamera(
+                CameraOptions.Builder()
+                    .center(currentUserLocation)
+                    .zoom(12.0)
+                    .build()
+            )
+            onUserCentered()
         }
     }
 }
